refactor: track_end_handler
