+++ src/events/mod.rs
@@ -1,2 +1,3 @@
 pub mod idle_handler;
 pub mod serenity_handler;
+pub mod track_end_handler;
